requirements.txt
fastapi
uvicorn
streamlit
pydantic
textblob
vaderSentiment
python-dotenv
requests
pandas
matplotlib
seaborn
google-generativeai